added null check

+++ van/server.py
@@ -151,8 +151,9 @@ def heartbeat():
             for received_id in json_back['received'][data_name]:
                 session.query(table).filter_by(id=received_id).delete()
 
-        for item in json_back['received']['heartbeat']:
-            session.query(Heartbeat).filter_by(id=item).delete()
+        if 'heartbeat' in json_back['recieved']:
+            for item in json_back['received']['heartbeat']:
+                session.query(Heartbeat).filter_by(id=item).delete()
 
         # Commit all the changes we made from the database
         session.commit()
